LayoutTests/fast/scrolling/ios/overflow-scroll-overlap-5.html

<!DOCTYPE html> <!-- webkit-test-runner [ internal:AsyncOverflowScrollingEnabled=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1">
<script src="../../../resources/ui-helper.js"></script>
<script src="../resources/overflow-scroll-overlap.js"></script>
<style>
.case {
    width: 200px;
    height: 200px;
    display: inline-block;
    position: relative;
}
.scrollcontent {
    width: 500px;
    height: 500px;
    background: green;
}

.overflowscroll {
    overflow: scroll;
    height: 100px;
    width: 100px;
    position: absolute;
    border: 2px solid black;
}
.overlapping {
    position:absolute;
    left: 25px;
    top: 25px;
    width: 100px;
    height: 100px;
    background: red;
}
.clip {
    position:absolute;
    width: 100px;
    height: 100px;
    overflow:hidden;
}
.large {
    width: 3000px;
    height: 150px;
}
#log {
    position:relative;
    white-space: pre;
}
</style>
</head>
<body onload="runTest()">
<p>
Test that scrollable areas with content that has complex containing block relationship are correctly targeted.
</p>

<div class="case">
    <div class="overflowscroll target">
        <div class="scrollcontent"></div>
        <div class="overlapping" style="position:absolute;"></div>
    </div>
</div>

<div class="case">
    <div class="overflowscroll target">
        <div class="scrollcontent"></div>
        <div class="overlapping" style="top:-475px; position:relative;"></div>
    </div>
</div>

<div class="case">
    <div class="overflowscroll target">
        <div class="scrollcontent"></div>
        <div class="overlapping" style="position:absolute;">
            <div class="overlapping" style="position:static; transform:translate3d(-50px, -5px, 0px)"></div>
        </div>
    </div>
</div>

<div class="case">
    <div class="overflowscroll target">
        <div class="scrollcontent"></div>
        <div class="overlapping" style="top:-475px; position:relative;">
            <div class="overlapping" style="position:static; transform:translate3d(-50px, -5px, 0px)"></div>
        </div>
    </div>
</div>

<div class="case">
    <div class="overflowscroll target">
        <div class="overflowscroll" style="left:25px; top:25px; position:absolute;">
            <div class="scrollcontent"></div>
        </div>
        <div class="scrollcontent"></div>
    </div>
</div>

<div class="case">
    <div class="overflowscroll target">
        <div class="scrollcontent"></div>
        <div class="overflowscroll" style="left:60px; top:60px; position:absolute;">
            <div class="scrollcontent"></div>
        </div>
    </div>
</div>

<div class="case">
    <div class="overflowscroll target">
        <div class="scrollcontent" style="position:absolute;"></div>
        <div class="overflowscroll" style="position:absolute; left:25px;">
            <div class="scrollcontent"></div>
            <div class="overlapping" style="position:absolute;">
                <div class="overlapping" style="position:static; transform:translate3d(-50px, -5px, 0px)"></div>
            </div>
        </div>
    </div>
</div>

<div class="case">
    <div class="overflowscroll target">
        <div class="scrollcontent" style="position:absolute;"></div>
        <div class="overflowscroll" style="position:absolute; left:60px;">
            <div class="scrollcontent"></div>
            <div class="overlapping" style="position:absolute;">
                <div class="overlapping" style="position:static; transform:translate3d(-50px, -5px, 0px)"></div>
            </div>
        </div>
    </div>
</div>

<div id=log></div>

</body>
</html>
